refactor: Move getRNSR to prepareRnsrJson
